Fixes the python venv errors
It seems that in the 5.0 version of the IDF, espressif has switched to
managing a python venv that the idf tries to install everything into.

Using the recommended commands actually works, but is redundant in our
case, because we have already used mach to construct a python
environment.

By setting IDF_PYTHON_ENV_PATH we are able to disable the IDF's attempt
to install a new virtual environment in $HOME/.espressif. Unfortunatly,
espressif seems to have rolled their own strange little lock file thingy
that is meant to validate all the installed versions of python packages.

The documentation notes that this is important, because without it the
idf may fail to function correctly. However, since we are super
smart, we can just blindly ignore that and assume that nixpkgs and mach
have it handled. They wouldn't betray us, right? right??

Anyway, we tell the idf not to try and check for python package versions
by also settings IDF_PYTHON_CHECK_CONSTRAINTS=no, and everything seems
to Just Work (tm).